Add 'show product details' functionality
Implemented navigation to the product description page (PDP) 
with the ability to persist state between page reload. 
Saving current product to the localStorage helps to fill up 
the PDP even if it was reloaded and thus the user won't 
have to navigate to homepage to retrieve the information 
on the product once again due to the data loss.