Move attributes validation process to the PDP
According to the requirements, the product can't be 
added to the cart without all of its attributes selected.
In the previous version the user could add the products to
the cart right from the PLP and the validation had happened
in the cart modal on navigation to the cart page.
In the current version the user has to select all attributes
even before adding the product to the cart, so the validation
has to happen on the PDP.