Highlight product attributes in the cart modal
When the cart modal opens, the previously selected 
attributes of all the products are being highlighted.
It also gives the ability to change the attributes before
proceeding to the cart page and all the changes are 
being saved to the state and localStorage so the user won't
have to select them once again after on page reload.